# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "COVID Modeling",
    "description": "The COVID Modeling UI was developed to help policy makers explore hypothetical intervention strategies to reduce the impact of COVID-19. It provides a common interface and language to run simulations against multiple models at the same time. This API is intended to facilitate automation, e.g. running weekly simulations with the same parameters, and integration with other tools.",
    "version": "0.0.1",
    "license": {"name": "MIT License", "url": "https://mit-license.org/"}
  },
  "servers": [{"url": "/api"}],
  "security": [{"token": [], "session": []}],
...
  "paths": {
    "/simulations": {
      "get": {
        "description": "Retrieves list of simulations",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/SimulationSummary"}
                }
              }
            }
          }
        },
        "operationId": "getSimulations",
        "tags": ["simulations"]
      },
      "post": {
        "description": "Schedule new simulation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/NewSimulationConfig"},
              "example": {
                "regionID": "US",
                "subregionID": "US-AK",
                "label": "Simulation",
                "customCalibrationDate": "2020-03-06",
                "interventionPeriods": [
                  {
                    "startDate": "2020-03-06",
                    "socialDistancing": "aggressive",
                    "schoolClosure": "aggressive",
                    "caseIsolation": "aggressive",
                    "voluntaryHomeQuarantine": "aggressive",
                    "reductionPopulationContact": 0
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "required": ["id"],
                  "properties": {"id": {"type": "integer"}}
                }
              }
            }
          },
          "422": {
            "description": "Invalid configuration",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {"error": {"type": "string"}}
                }
              }
            }
          }
        },
        "operationId": "postSimulations",
        "tags": ["simulations"]
      }
    },
...
    "/simulations/{id}": {
      "get": {
        "description": "Retrieves simulation summary",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Simulation ID",
            "schema": {"type": "integer"},
            "example": 84
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {"$ref": "#/components/schemas/Simulation"}
              }
            }
          }
        },
        "operationId": "getSimulation",
        "tags": ["simulations"]
      }
    },
...
    "/simulations/{id}/model-runs/{model}/download": {
      "get": {
        "description": "Downloads full result of simulation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Simulation ID",
            "schema": {"type": "integer"},
            "example": 2
          },
          {
            "in": "path",
            "name": "model",
            "description": "Model slug",
            "schema": {
              "type": "string",
              "$ref": "#/components/schemas/ModelSlug"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Successful redirect",
            "headers": {
              "Location": {
                "description": "The URL to download results from",
                "schema": {"type": "string"}
              }
            }
          }
        },
        "operationId": "getSimulationDownload",
        "tags": ["model-runs"]
      }
    },
    "/simulations/{id}/model-runs/{model}/export": {
      "get": {
        "description": "Downloads full result of simulation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Simulation ID",
            "schema": {"type": "integer"},
            "example": 84
          },
          {
            "in": "query",
            "name": "path",
            "description": "Model slug",
            "schema": {
              "type": "string",
              "$ref": "#/components/schemas/ModelSlug"
            }
          },
          {
            "in": "query",
            "name": "format",
            "description": "Determines what data and media type to return:\n \n * results -> application/json\n * crystalcast -> text/csv",
            "default": "results",
            "schema": {
              "type": "string",
              "$ref": "#/components/schemas/ExportFormat"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful export",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {"$ref": "#/components/schemas/ModelOutput"}
              },
              "text/csv": {
                "schema": {"type": "string"},
                "examples": {
                  "success": {"externalValue": "/export-example.csv"}
                }
              }
            }
          }
        },
        "operationId": "getSimulationExport",
        "tags": ["model-runs"]
      }
    },
    "/simulations/{id}/model-runs/{model}": {
      "get": {
        "description": "Retrieves model run summary",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Simulation ID",
            "schema": {"type": "integer"},
            "example": 84
          },
          {
            "in": "path",
            "name": "model",
            "description": "Model slug",
            "schema": {
              "type": "string",
              "$ref": "#/components/schemas/ModelSlug"
            },
            "example": "mrc-ide-covid-sim"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {"$ref": "#/components/schemas/ModelRun"}
              }
            }
          }
        },
        "operationId": "getSimulation",
        "tags": ["model-runs"]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExportFormat": {"type": "string", "enum": ["results", "crystalcast"]},
      "Intensity": {
        "type": "string",
        "enum": ["aggressive", "moderate", "mild"],
        "title": "Intensity"
      },
      "InterventionPeriod": {
        "type": "object",
        "properties": {
          "startDate": {
            "$ref": "#/components/schemas/ISODate",
            "description": "An ISO-8601 string encoding the date that these interventions begin."
          },
          "socialDistancing": {
            "$ref": "#/components/schemas/Intensity",
            "description": "The level of social distancing in the region."
          },
          "schoolClosure": {
            "$ref": "#/components/schemas/Intensity",
            "description": "The level of school closure in the region."
          },
          "caseIsolation": {
            "$ref": "#/components/schemas/Intensity",
            "description": "The level to which individuals with symptoms self-isolate."
          },
          "voluntaryHomeQuarantine": {
            "$ref": "#/components/schemas/Intensity",
            "description": "The level to which entire households self-isolate when one member of the household has symptoms."
          },
          "reductionPopulationContact": {
            "type": "number",
            "description": "The estimated reduction in population contact resulting from all of the above interventions.\nSome models require this generalized parameter instead of the individual interventions."
          }
        },
        "required": ["startDate", "reductionPopulationContact"],
        "additionalProperties": false,
        "description": "Each intervention period is specified by a `startDate`, a set of interventions (`socialDistancing`, `caseIsolation`, `voluntaryHomeQuarantine`, and `schoolClosure`), and an estimate of the overall effect of these interventions (`reductionPopulationContact`).\nThis overall estimate is needed because some models do not simulate the effects of individual interventions.\nThe strictness of each intervention is specified roughly, as one of `mild`, `moderate`, or `aggressive`.\nEach model connector is responsible for interpreting this distinction in a way that works for the particular model.",
        "title": "Intervention Period"
      },
      "ISODate": {
        "type": "string",
        "title": "ISO Date",
        "example": "2020-03-06"
      },
      "ModelDescription": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "A short display name to identify the model"
          },
          "modelVersion": {
            "type": "string",
            "description": "A version number identifying the version of the model used in this run."
          },
          "connectorVersion": {
            "type": "string",
            "description": "A version number identifying the version of the connector used in this run.\nThis may be the same as modelVersion if development of the two is closely-linked."
          }
        },
        "required": ["name", "modelVersion", "connectorVersion"],
        "additionalProperties": false,
        "title": "Model Description"
      },
      "CommonModelInput": {
        "type": "object",
        "properties": {
          "region": {
            "$ref": "#/components/schemas/Region",
            "description": "The desired geographical top-level region for the simulation.\nConnector must check this value (and subregion, if provided) and and either produce results for that geography, or return an error."
          },
          "subregion": {
            "$ref": "#/components/schemas/Subregion",
            "description": "Specific subregion for the simulation.\nMay be omitted"
          },
          "parameters": {
            "$ref": "#/components/schemas/ModelParameters",
            "description": "Desired parameters.\nConnectors may choose to ignore any of these parameters if they wish."
          }
        },
        "required": ["region", "parameters"],
        "additionalProperties": false,
        "description": "A generalized description of the input to an epidemiological model.",
        "title": "Common Model Input"
      },
...
      "MinimalModelInput": {
        "type": "object",
        "properties": {
          "p": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Vector of model parameters"
          },
          "u0": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Vector of initial conditions"
          },
          "tspan": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Timespan over which to simulate the model"
          },
          "dt": {
            "type": "number",
            "description": "Optional time interval at which to compute the states."
          }
        },
        "required": ["p", "u0", "tspan"],
        "additionalProperties": false,
        "title": "Minimal Model Input"
      },
      "ModelInput": {
        "anyOf": [
          {"$ref": "#/components/schemas/CommonModelInput"},
          {"$ref": "#/components/schemas/MinimalModelInput"},
          {"$ref": "#/components/schemas/MicroMoBBHRMInput"}
        ],
        "title": "Model Input"
      },
      "CommonModelOutput": {
        "type": "object",
        "properties": {
          "metadata": {"$ref": "#/components/schemas/CommonModelInput"},
          "model": {"$ref": "#/components/schemas/ModelDescription"},
          "time": {
            "type": "object",
            "properties": {
              "t0": {
                "type": "string",
                "description": "An ISO-8601 string encoding the date that each timeseries begins."
              },
              "timestamps": {
                "type": "array",
                "items": {"type": "number"},
                "description": "The timestamps that correspond to every series of metrics.\nEach value is a number of days after `t0`."
              },
              "extent": {
                "type": "array",
                "minItems": 2,
                "items": [{"type": "number"}, {"type": "number"}],
                "maxItems": 2,
                "description": "The minimum and maximum timestamps for the series of reported metrics.\nEach value is a number of days after `t0`."
              }
            },
            "required": ["t0", "timestamps", "extent"],
            "additionalProperties": false
          },
          "aggregate": {
            "type": "object",
            "properties": {
              "metrics": {"$ref": "#/components/schemas/SeverityMetrics"}
            },
            "required": ["metrics"],
            "additionalProperties": false
          }
        },
        "required": ["metadata", "model", "time", "aggregate"],
        "additionalProperties": false,
        "title": "Common Model Output"
      },
      "MicroMoBBHRMOutput": {
        "type": "object",
        "properties": {
          "metadata": {"$ref": "#/components/schemas/MicroMoBBHRMInput"},
          "t": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Vector of times at which the model is run"
          },
          "MYZ": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {"type": "array", "items": {"type": "number"}}
            },
            "description": "Array (patches, days, MYZ)"
          }
        },
        "required": ["metadata", "t", "MYZ"],
        "additionalProperties": false,
        "title": "Output for MicroMoB model with Beverton-Holt aquatic mosquito model and\nRoss-Macdonald adult mosquito model."
      },
      "MinimalModelOutput": {
        "type": "object",
        "properties": {
          "metadata": {"$ref": "#/components/schemas/MinimalModelInput"},
          "model": {"$ref": "#/components/schemas/ModelDescription"},
          "t": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Vector of times at which the model is run"
          },
          "u": {
            "type": "array",
            "items": {"type": "array", "items": {"type": "number"}},
            "description": "Matrix of states"
          },
          "outputs": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Optional vector of outputs"
          }
        },
        "required": ["metadata", "model", "t", "u"],
        "additionalProperties": false,
        "title": "Minimal Model Output"
      },
      "ModelOutput": {
        "anyOf": [
          {"$ref": "#/components/schemas/CommonModelOutput"},
          {"$ref": "#/components/schemas/MinimalModelOutput"},
          {"$ref": "#/components/schemas/MicroMoBBHRMOutput"}
        ],
        "title": "Model Output"
      },
      "ModelParameters": {
        "type": "object",
        "properties": {
          "calibrationDate": {
            "$ref": "#/components/schemas/ISODate",
            "description": "An ISO-8601 string encoding the date of the most recent case data in the region.\nSome models such as `CovidSim` require you to provide some information about recorded cases in the region, in order to calibrate the model.\nThe web-interface provides this information automatically, based on recorded case data downloaded daily from various international data sources."
          },
          "calibrationCaseCount": {
            "type": "number",
            "description": "The total number of confirmed cases in the region before the calibration date."
          },
          "calibrationDeathCount": {
            "type": "number",
            "description": "The total number of deaths in the region before the calibration date."
          },
          "r0": {
            "type": "number",
            "description": "The assumed [_reproduction number_](https://en.wikipedia.org/wiki/Basic_reproduction_number) for the virus.\nIf this is null, then each model will use its own default value.",
            "nullable": true
          },
          "interventionPeriods": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/InterventionPeriod"},
            "description": "A list of time periods, each with a different set of interventions.\nPolicy interventions are specified as a series of _intervention periods_, each with a certain set of interventions that are in place.\nFor example, case isolation and social distancing may be instituted first, followed by school closure a week later, followed by a relaxation of all guidelines after several months.\n**Note** - In order to specify that _all_ interventions end on a certain date, there should be a _final_ intervention period that starts on that date, has no interventions specified, and has `reductionPopulationContact` set to zero."
          }
        },
        "required": [
          "calibrationDate",
          "calibrationCaseCount",
          "calibrationDeathCount",
          "r0",
          "interventionPeriods"
        ],
        "additionalProperties": false,
        "title": "Model Parameters"
      },
      "ModelRun": {
        "type": "object",
        "properties": {
          "model_slug": {"$ref": "#/components/schemas/ModelSlug"},
          "status": {"$ref": "#/components/schemas/RunStatus"},
          "results_data": {"type": "string", "nullable": true},
          "export_location": {"type": "string", "nullable": true}
        },
        "required": ["model_slug", "status", "results_data", "export_location"],
        "additionalProperties": false
      },
      "ModelSlug": {"type": "string"},
...
      "NewSimulationConfig": {
        "type": "object",
        "properties": {
          "regionID": {"$ref": "#/components/schemas/Region"},
          "subregionID": {"$ref": "#/components/schemas/Subregion"},
          "label": {"type": "string"},
          "r0": {"type": "number", "example": 1},
          "customCalibrationDate": {"$ref": "#/components/schemas/ISODate"},
          "interventionPeriods": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/InterventionPeriod"}
          }
        },
        "required": ["regionID", "label", "interventionPeriods"],
        "additionalProperties": false,
        "description": "The data sent from web UI to backend"
      },
      "Region": {"type": "string", "title": "Region", "example": "US"},
      "RunStatus": {
        "type": "string",
        "enum": ["pending", "in-progress", "complete", "unsupported", "failed"],
        "title": "Run Status"
      },
...
      "Simulation": {
        "type": "object",
        "properties": {
          "id": {"type": "number"},
          "region_name": {"type": "string"},
          "status": {"$ref": "#/components/schemas/RunStatus"},
          "subregion_name": {"type": "string"},
          "region_id": {"type": "string"},
          "subregion_id": {"type": "string"},
          "github_user_id": {"type": "number"},
          "github_user_login": {"type": "string"},
          "configuration": {"$ref": "#/components/schemas/ModelInput"},
          "model_runs": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ModelRun"}
          },
          "label": {"type": "string"},
          "created_at": {"type": "string"},
          "updated_at": {"type": "string"}
        },
        "required": [
          "id",
          "region_name",
          "status",
          "github_user_id",
          "github_user_login",
          "configuration",
          "model_runs",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "SimulationSummary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "configurationSummary": {"type": "string"},
          "id": {"type": "number"},
          "region_name": {"type": "string"},
          "status": {"$ref": "#/components/schemas/RunStatus"},
          "subregion_name": {"type": "string"},
          "region_id": {"type": "string"},
          "subregion_id": {"type": "string"},
          "github_user_id": {"type": "number"},
          "github_user_login": {"type": "string"},
          "model_runs": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ModelRun"}
          },
          "label": {"type": "string"},
          "created_at": {"type": "string"},
          "updated_at": {"type": "string"}
        },
        "required": [
          "configurationSummary",
          "created_at",
          "github_user_id",
          "github_user_login",
          "id",
          "model_runs",
          "region_name",
          "status",
          "updated_at"
        ]
      },
      "Subregion": {"type": "string", "title": "Subregion", "example": "US-AK"}
    },
    "securitySchemes": {
      "token": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Primary authentication method for programmatic use."
      },
      "session": {
        "type": "http",
        "in": "cookie",
        "name": "session",
        "description": "For use in UI. Also requires valid session.sig cookie"
      }
    }
  }
}